removing box and combobox files and making new Landingcomponenets file for all the components
